Adapt the status bar to only call GCP if needed.
Previously we call GCP with a subprocess each seconds. This behaviour induces
network overload and subprocess spawning.
Now, we call the GCP only if a new commit was pushed for the current branch
since the last call, or if the last call returns a job in status RUNNING or
PENDING.